backend/requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
flask-cors==3.0.10
elasticsearch==7.17.0
python-dotenv==0.19.0
requests==2.26.0
elasticsearch-dsl==7.4.0
urllib3==1.26.6 
pyjwt
openai